+++ src/main/java/kitchenpos/menu/dto/MenuRequestDto.java
@@ -1,8 +1,8 @@
 package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu_group.domain.MenuGroup;
 
 import java.math.BigDecimal;
 import java.util.List;
