+++ src/test/java/kitchenpos/common/fixture/MenuFixture.java
@@ -1,12 +1,12 @@
 package kitchenpos.common.fixture;
 
 import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.dto.MenuProductRequestDto;
 import kitchenpos.menu.dto.MenuProductResponseDto;
 import kitchenpos.menu.dto.MenuRequestDto;
 import kitchenpos.menu.dto.MenuResponseDto;
-import kitchenpos.menu_group.domain.MenuGroup;
 
 import java.math.BigDecimal;
 import java.util.List;
