+++ src/test/java/kitchenpos/order/repository/OrderRepositoryTest.java
@@ -1,7 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.order.repository;
 
 import kitchenpos.common.domain.OrderStatus;
-import kitchenpos.order.repository.OrderRepository;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
