+++ src/test/java/kitchenpos/table/repository/OrderTableRepositoryTest.java
@@ -1,7 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.table.repository;
 
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.repository.OrderTableRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
